# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Geek Social API",
    "version": "1.0.0",
    "description": "API REST do Geek Social — rede social para gamers e colecionadores."
  },
  "components": {
    "securitySchemes": {
      "accessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token enviado no header Authorization. Obtido via /auth/login."
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "HttpOnly cookie usado em /auth/refresh."
      }
    },
...
  },
  "paths": {
...
    "/integrations/steam/login": {
      "post": {
        "operationId": "steam_login",
        "summary": "Iniciar vínculo Steam (OpenID)",
        "tags": [
          "Steam"
        ],
        "description": "Retorna URL de autenticação OpenID. Frontend redireciona usuário pra ela. Callback faz o link.",
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/integrations/steam/link": {
      "delete": {
        "operationId": "steam_unlink",
        "summary": "Desvincular Steam",
        "tags": [
          "Steam"
        ],
        "description": "Limpa users.steam_id, steam_linked_at e steam_api_key. Itens já importados ficam preservados nas coleções.",
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/integrations/steam/api-key": {
      "put": {
        "operationId": "steam_set_api_key",
        "summary": "Definir Steam Web API key do user",
        "tags": [
          "Steam"
        ],
        "description": "Necessária pra importação (api.steampowered.com não permite consultas anônimas). User obtém em https://steamcommunity.com/dev/apikey. 32 chars hex.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "apiKey": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 32
                  }
                },
                "required": [
                  "apiKey"
                ]
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "operationId": "steam_clear_api_key",
        "summary": "Limpar API key salva",
        "tags": [
          "Steam"
        ],
        "description": "Remove steam_api_key. Importação fica indisponível até nova key ser setada.",
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/integrations/steam/games": {
      "get": {
        "operationId": "steam_list_games",
        "summary": "Listar jogos owned pelo usuário no Steam",
        "tags": [
          "Steam"
        ],
        "description": "Chama GetOwnedGames da Steam Web API com a key salva. Resposta inclui appId, name, playtimeForever, iconUrl. Falha com STEAM_PROFILE_PRIVATE se perfil não público.",
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/integrations/steam/import": {
      "post": {
        "operationId": "steam_import_start",
        "summary": "Iniciar importação de jogos selecionados",
        "tags": [
          "Steam"
        ],
        "description": "Recebe appIds + collection destino (existente ou nova). Cria batch_id, enfileira jobs steam.import-game (1 por appId), e enrich-game subsequente. Retorna { batchId } pra polling.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "collectionId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "newCollectionName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "appIds": {
                    "minItems": 1,
                    "maxItems": 2000,
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "gamesSnapshot": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "appId": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        },
                        "name": {
                          "type": "string"
                        },
                        "playtimeForever": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "appId",
                        "name",
                        "playtimeForever"
                      ]
                    }
                  }
                },
                "required": [
                  "appIds"
                ]
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/integrations/steam/import/{batchId}/status": {
      "get": {
        "operationId": "steam_import_status",
        "summary": "Status da importação",
        "tags": [
          "Steam"
        ],
        "description": "Polling: retorna { total, processed, failed, finalized, finalizedAt? }. Frontend usa pra progress bar — recomenda-se também observar socket events import:progress / import:done.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "batchId",
            "required": true
          }
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3003",
      "description": "Dev local"
    }
  ],
...
}